Revert "Prepare for v2"
This reverts commit 64276bd8817b92577b36d7e8fa051f067c3b3137.
