+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cncjs-controller",
-  "version": "2.0.0",
+  "version": "1.3.0",
   "description": "A controller library for event-based communication between client and CNCjs server.",
   "main": "lib/index.js",
   "scripts": {
@@ -33,9 +33,7 @@
     "events",
     "client"
   ],
-  "dependencies": {
-    "lodash.mapvalues": "^4.6.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "babel-cli": "~6.24.1",
     "babel-core": "~6.24.1",
