--- src/units.js
@@ -1,10 +1,0 @@
-// from mm to in
-const mm2in = (val = 0) => val / 25.4;
-
-// from in to mm
-const in2mm = (val = 0) => val * 25.4;
-
-export {
-    mm2in,
-    in2mm
-};
